(PC-13940)[PRO] feat: move no offers below filters
